Moved include layer processor to its own file
